Move unit tests into distribution sub-submodule
